+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## `main`
 * Make `extendedFields` default to empty instead of `nil`.
 * Use `Version` type to represent SemVers instead of Strings.
+* Make types `Sendable`.
 
 ## 0.2.0
 * Fix `TileJSON` init parameters